Fix connection non-reuse issue due to clobbered cipher_list
This commit makes a clone of the connection's ssl config
cipher_list before passing it to nss.c's set_ciphers method
as that method clobbers its input by changing commas to
nulls.

Without this commit if a multiple entry cipher list is specified
than reuse of a connection in the connection cache is not possible.

This addresses https://github.com/curl/curl/issues/8160.